Minor restructuring for schema definition validation
